P13 Made shared component for add and edit expense screens.
